requirements/base.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.26.5

# Filtering & CORS
django-filter==23.4
django-cors-headers==4.3.0

# Environment variables
python-dotenv==1.0.0

# WSGI server for production
gunicorn==21.2.0
